[feat] TransferController 관련 DTO 그루핑
