+++ src/main/kotlin/com/chatandpay/api/controller/TransferController.kt
@@ -16,7 +16,7 @@ class TransferController (
 
 
     @PostMapping("")
-    fun sendTransfer(@RequestBody request: ReceiveTransferRequestDTO): ResponseEntity<ReceiveTransferResponseDTO> {
+    fun sendTransfer(@RequestBody request: TransferDTO.ReceiveTransferRequestDTO): ResponseEntity<TransferDTO.ReceiveTransferResponseDTO> {
 
         val response = redissonLockFacade.sendTransfer(request)
         return ResponseEntity.ok(response)
@@ -25,7 +25,7 @@ class TransferController (
 
 
     @PostMapping("/{transfer_id}")
-    fun receiveTransfer(@PathVariable("transfer_id") id: UUID): ResponseEntity<SendTransferResponseDTO> {
+    fun receiveTransfer(@PathVariable("transfer_id") id: UUID): ResponseEntity<TransferDTO.SendTransferResponseDTO> {
 
         val response = transferService.receiveTransfer(id)
         return ResponseEntity.ok(response)
@@ -33,7 +33,7 @@ class TransferController (
     }
 
     @PostMapping("/interbank")
-    fun sendOtherBankTransfer(@RequestBody request: OtherBankTransferRequestDTO): ResponseEntity<OtherBankTransferResponseDTO> {
+    fun sendOtherBankTransfer(@RequestBody request: TransferDTO.OtherBankTransferRequestDTO): ResponseEntity<TransferDTO.OtherBankTransferResponseDTO> {
 
         val response = transferService.sendOtherBankTransfer(request)
         return ResponseEntity.ok(response)
@@ -41,7 +41,7 @@ class TransferController (
     }
 
     @PostMapping("/registered")
-    fun sendMyOtherBankTransfer(@RequestBody request: RegOtherBankTransferRequestDTO): ResponseEntity<OtherBankTransferResponseDTO> {
+    fun sendMyOtherBankTransfer(@RequestBody request: TransferDTO.RegOtherBankTransferRequestDTO): ResponseEntity<TransferDTO.OtherBankTransferResponseDTO> {
 
         val response = transferService.sendMyOtherBankTransfer(request)
         return ResponseEntity.ok(response)
@@ -49,7 +49,7 @@ class TransferController (
     }
 
     @GetMapping("/pending/{sender}")
-    fun getPendingTransfers(@PathVariable("sender") id: Long) : ResponseEntity<List<PendingTransferDTO>> {
+    fun getPendingTransfers(@PathVariable("sender") id: Long) : ResponseEntity<List<TransferDTO.PendingTransferDTO>> {
 
         // TODO 로그인 유지 (세션 등) 구현 후 sender부 변경
         val response = transferService.getPendingTransfers(id)
@@ -59,7 +59,7 @@ class TransferController (
 
 
     @PostMapping("/change")
-    fun changePendingTransferState(@RequestBody request: ChangePendingTransferRequestDTO) : ResponseEntity<ChangePendingTransferResponseDTO> {
+    fun changePendingTransferState(@RequestBody request: TransferDTO.ChangePendingTransferRequestDTO) : ResponseEntity<TransferDTO.ChangePendingTransferResponseDTO> {
 
         val response = transferService.changePendingTransferState(request)
         return ResponseEntity.ok(response)
