--- src/main/kotlin/com/chatandpay/api/dto/PendingTransferDTO.kt
@@ -1,12 +1,0 @@
-package com.chatandpay.api.dto
-
-import java.util.*
-
-data class PendingTransferDTO (
-
-    val uuid: UUID,
-    val sender: Long,
-    val receiver: Long,
-    val amount: Int
-
-)
