--- src/main/kotlin/com/chatandpay/api/dto/ReceiveTransferRequestDTO.kt
@@ -1,15 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.common.ApiResponse
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class ReceiveTransferRequestDTO(
-
-    val senderId: Long,
-    val receiverId: Long,
-    val amount: Int
-
-): ApiResponse()
