--- src/main/kotlin/com/chatandpay/api/dto/SendTransferResponseDTO.kt
@@ -1,15 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.common.ApiResponse
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-import java.util.*
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class SendTransferResponseDTO (
-  
-    val transferUuid: UUID,
-    val receivedAmount: Int,
-    val walletAmount: Int,
-
-) : ApiResponse()
